don't use overrideScalaVersion / autoBoot / filterLibrary for 2.13 and 3
`overrideScalaVersion` is part of ScalaModuleInfo in sbt. If true
and also `autoScalaLibrary` is true, coursier's `autoScalaLibrary`
will be set to true. This is passed forward to coursier's
`forceScalaVersion` => `doForceScalaVersion`, which sets consistent
versions for library/reflect/compiler.

`autoBoot` is part of ClasspathOptions in zinc's compiler interface.
If true the compiler will be invoked with `-bootclasspath scala-library`.

`filterLibrary` is part of ClasspathOptions in zinc's compiler interface.
If true, the scala-library is removed from the classpath passed to
the compiler's `-cp`.